+++ src/App.vue
@@ -69,7 +69,8 @@
               :move-duration-ms="timing.moveMs"
               :move-easing="timing.moveEasing"
               :board-sound-callbacks="boardSoundCallbacks"
-              :on-sound-unlock="activateSounds"
+              :on-audio-unlock="unlockSoundsSync"
+              :on-audio-warm-up="warmUpSounds"
               @started="onGameStarted"
               @ended="onGameEnded"
               @score="onGameScore"
@@ -180,7 +181,8 @@ const {
   onGameAimReached,
   onSessionUpdate,
   boardSoundCallbacks,
-  activateSounds,
+  unlockSoundsSync,
+  warmUpSounds,
 } = useGameController(gameContainerEl)
 </script>